MOSIP-44959 fixed test file
